Prefer conda over pip, and pip over apt
The goal of this change is to reduce the need for root
privileges to install software to support PANOPTES.
It is also an experiment with making more use of conda,
for which we could conceivably create our own packages
and upload them to the anaconda.org repository system.

If this goes well, then there are more packages that
we could shift from apt to conda (e.g. ffmpeg). There
are comments in apt-packages-list.txt about some
candidates for that shift.